+++ backend/modules/extensions/engine/model.php
@@ -366,8 +366,8 @@ class BackendExtensionsModel
 			// remove items that are not for the current language
 			if(isset($row['data']['language']) && $row['data']['language'] != BackendLanguage::getWorkingLanguage()) $itemsToRemove[] = $id;
 
-			// set URL if needed
-			if(!isset($row['data']['url'])) $row['data']['url'] = BackendModel::createURLForAction('index', $row['module']);
+			// set URL if needed, we use '' instead of null, because otherwise the module of the current action (modules) is used.
+			if(!isset($row['data']['url'])) $row['data']['url'] = BackendModel::createURLForAction('', $row['module']);
 
 			// build name
 			$name = SpoonFilter::ucfirst(BL::lbl($row['label']));