+++ frontend/core/engine/base.php
@@ -472,6 +472,9 @@ class FrontendBaseBlock
 		$pagination['num_pages'] = $this->pagination['num_pages'];
 		$pagination['current_page'] = $this->pagination['requested_page'];
 
+		// define anchor
+		$anchor = (isset($this->pagination['anchor'])) ? '#' . $this->pagination['anchor'] : '';
+
 		// as long as we are below page 5 we should show all pages starting from 1
 		if($this->pagination['requested_page'] <= 6)
 		{
@@ -513,13 +516,13 @@ class FrontendBaseBlock
 
 			// set
 			$pagination['show_previous'] = true;
-			$pagination['previous_url'] = $URL;
+			$pagination['previous_url'] = $URL . $anchor;
 
 			// flip ahead
 			$this->header->addLink(
 				array(
 				     'rel' => 'prev',
-				     'href' => SITE_URL . $URL,
+				     'href' => SITE_URL . $URL . $anchor,
 				)
 			);
 		}
@@ -539,7 +542,7 @@ class FrontendBaseBlock
 				else $URL = $this->pagination['url'] . '&amp;page=' . $i;
 
 				// add
-				$pagination['first'][] = array('url' => $URL, 'label' => $i);
+				$pagination['first'][] = array('url' => $URL . $anchor, 'label' => $i);
 			}
 		}
 
@@ -554,7 +557,7 @@ class FrontendBaseBlock
 			else $URL = $this->pagination['url'] . '&amp;page=' . $i;
 
 			// add
-			$pagination['pages'][] = array('url' => $URL, 'label' => $i, 'current' => $current);
+			$pagination['pages'][] = array('url' => $URL . $anchor, 'label' => $i, 'current' => $current);
 		}
 
 		// show last pages?
@@ -572,7 +575,7 @@ class FrontendBaseBlock
 				else $URL = $this->pagination['url'] . '&amp;page=' . $i;
 
 				// add
-				$pagination['last'][] = array('url' => $URL, 'label' => $i);
+				$pagination['last'][] = array('url' => $URL . $anchor, 'label' => $i);
 			}
 		}
 
@@ -585,13 +588,13 @@ class FrontendBaseBlock
 
 			// set
 			$pagination['show_next'] = true;
-			$pagination['next_url'] = $URL;
+			$pagination['next_url'] = $URL . $anchor;
 
 			// flip ahead
 			$this->header->addLink(
 				array(
 				     'rel' => 'next',
-				     'href' => SITE_URL . $URL,
+				     'href' => SITE_URL . $URL . $anchor,
 				)
 			);
 		}