+++ frontend/modules/blog/engine/model.php
@@ -73,7 +73,7 @@ class FrontendBlogModel implements FrontendTagsInterface
 		$items = (array) FrontendModel::getDB()->getRecords(
 			'SELECT i.id, i.revision_id, i.language, i.title, i.introduction, i.text, i.num_comments AS comments_count,
 			 c.title AS category_title, m2.url AS category_url, i.image,
-			 UNIX_TIMESTAMP(i.publish_on) AS publish_on, i.user_id,
+			 UNIX_TIMESTAMP(i.publish_on) AS publish_on, i.user_id, i.allow_comments,
 			 m.url
 			 FROM blog_posts AS i
 			 INNER JOIN blog_categories AS c ON i.category_id = c.id
@@ -103,7 +103,13 @@ class FrontendBlogModel implements FrontendTagsInterface
 			// comments
 			if($row['comments_count'] > 0) $items[$key]['comments'] = true;
 			if($row['comments_count'] > 1) $items[$key]['comments_multiple'] = true;
-
+			
+			// allow comments as boolean
+			$items[$key]['allow_comments'] = ($row['allow_comments'] == 'Y');
+			
+			// reset allow comments
+			if(!FrontendModel::getModuleSetting('blog', 'allow_comments')) $items[$key]['allow_comments'] = false;
+			
 			// image?
 			if(isset($row['image']))
 			{
@@ -201,7 +207,7 @@ class FrontendBlogModel implements FrontendTagsInterface
 		$items = (array) FrontendModel::getDB()->getRecords(
 			'SELECT i.id, i.revision_id, i.language, i.title, i.introduction, i.text, i.num_comments AS comments_count,
 			 c.title AS category_title, m2.url AS category_url, i.image,
-			 UNIX_TIMESTAMP(i.publish_on) AS publish_on, i.user_id,
+			 UNIX_TIMESTAMP(i.publish_on) AS publish_on, i.user_id, i.allow_comments,
 			 m.url
 			 FROM blog_posts AS i
 			 INNER JOIN blog_categories AS c ON i.category_id = c.id
@@ -231,6 +237,12 @@ class FrontendBlogModel implements FrontendTagsInterface
 			// comments
 			if($row['comments_count'] > 0) $items[$key]['comments'] = true;
 			if($row['comments_count'] > 1) $items[$key]['comments_multiple'] = true;
+
+			// allow comments as boolean
+			$items[$key]['allow_comments'] = ($row['allow_comments'] == 'Y');
+
+			// reset allow comments
+			if(!FrontendModel::getModuleSetting('blog', 'allow_comments')) $items[$key]['allow_comments'] = false;
 
 			// image?
 			if(isset($row['image']))
@@ -287,7 +299,7 @@ class FrontendBlogModel implements FrontendTagsInterface
 		$items = (array) FrontendModel::getDB()->getRecords(
 			'SELECT i.id, i.revision_id, i.language, i.title, i.introduction, i.text, i.num_comments AS comments_count,
 			 c.title AS category_title, m2.url AS category_url, i.image,
-			 UNIX_TIMESTAMP(i.publish_on) AS publish_on, i.user_id,
+			 UNIX_TIMESTAMP(i.publish_on) AS publish_on, i.user_id, i.allow_comments,
 			 m.url
 			 FROM blog_posts AS i
 			 INNER JOIN blog_categories AS c ON i.category_id = c.id
@@ -315,6 +327,12 @@ class FrontendBlogModel implements FrontendTagsInterface
 			// comments
 			if($row['comments_count'] > 0) $items[$key]['comments'] = true;
 			if($row['comments_count'] > 1) $items[$key]['comments_multiple'] = true;
+
+			// allow comments as boolean
+			$items[$key]['allow_comments'] = ($row['allow_comments'] == 'Y');
+
+			// reset allow comments
+			if(!FrontendModel::getModuleSetting('blog', 'allow_comments')) $items[$key]['allow_comments'] = false;
 
 			// image?
 			if(isset($row['image']))
@@ -548,10 +566,10 @@ class FrontendBlogModel implements FrontendTagsInterface
 			'SELECT i.id, i.title, CONCAT(?, m.url) AS url
 			 FROM blog_posts AS i
 			 INNER JOIN meta AS m ON i.meta_id = m.id
-			 WHERE i.id != ? AND i.status = ? AND i.hidden = ? AND i.language = ? AND i.publish_on <= ?
-			 ORDER BY i.publish_on DESC
+			 WHERE i.id != ? AND i.status = ? AND i.hidden = ? AND i.language = ? AND ((i.publish_on = ? AND i.id < ?) OR i.publish_on < ?)
+			 ORDER BY i.publish_on DESC, i.id DESC
 			 LIMIT 1',
-			array($detailLink, $id, 'active', 'N', FRONTEND_LANGUAGE, $date)
+			array($detailLink, $id, 'active', 'N', FRONTEND_LANGUAGE, $date, $id, $date)
 		);
 
 		// get next post
@@ -559,10 +577,10 @@ class FrontendBlogModel implements FrontendTagsInterface
 			'SELECT i.id, i.title, CONCAT(?, m.url) AS url
 			 FROM blog_posts AS i
 			 INNER JOIN meta AS m ON i.meta_id = m.id
-			 WHERE i.id != ? AND i.status = ? AND i.hidden = ? AND i.language = ? AND i.publish_on > ?
-			 ORDER BY i.publish_on ASC
+			 WHERE i.id != ? AND i.status = ? AND i.hidden = ? AND i.language = ? AND ((i.publish_on = ? AND i.id > ?) OR i.publish_on > ?)
+			 ORDER BY i.publish_on ASC, i.id ASC
 			 LIMIT 1',
-			array($detailLink, $id, 'active', 'N', FRONTEND_LANGUAGE, $date)
+			array($detailLink, $id, 'active', 'N', FRONTEND_LANGUAGE, $date, $id, $date)
 		);
 
 		return $navigation;